style: fix unused variables and simplify code patterns
- Prefix unused variables with underscore (RUF059, F841)
- Simplify exception handling with contextlib.suppress (SIM105)
- Combine nested if statements (SIM102)
- Simplify boolean comparisons (E712)
- Remove needless bool conversions (SIM103)
- Rename unused loop variables (B007)
- Reduces ruff issues from 1049 to 240 (77% reduction)

Total reduction this session: 1830 → 240 (87% reduction)

+++ ai-service/scripts/population_based_training.py
@@ -137,7 +137,7 @@ def sample_hyperparameter(name: str, current: float | None = None, perturb: bool
             new_val = np.random.uniform(min_val, max_val)
 
     if is_int:
-        new_val = int(round(new_val))
+        new_val = round(new_val)
 
     return new_val
